# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "Audit": {
        "properties": {
          "chain": {
            "$ref": "#/components/schemas/ChainStatus"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "required": [
          "chain",
          "entries"
        ],
        "title": "Audit",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RollbackRequest": {
        "additionalProperties": false,
        "properties": {
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Version to pin; null unpins",
            "title": "Version"
          }
        },
        "title": "RollbackRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Token from POST /v1/auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Registers teams, watches the pool, and controls the gateways.",
    "title": "Resilient IP Optimizer control API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/audit": {
      "get": {
        "operationId": "get_audit_v1_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Audit"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Role too low"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Audit",
        "tags": [
          "audit"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Wrong email or password"
          },
          "403": {
            "description": "Role too low"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/config/rollback": {
      "post": {
        "operationId": "config_rollback_v1_config_rollback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rollback"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Role too low"
          },
          "404": {
            "description": "No such version"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Config Rollback",
        "tags": [
          "config"
        ]
      }
    },
    "/v1/config/versions": {
      "get": {
        "operationId": "config_versions_v1_config_versions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigVersions"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Role too low"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Config Versions",
        "tags": [
          "config"
        ]
      }
    },
    "/v1/config/versions/{version}": {
      "get": {
        "operationId": "config_version_v1_config_versions__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigVersionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Role too low"
          },
          "404": {
            "description": "No such version"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Config Version",
        "tags": [
          "config"
        ]
      }
    },
    "/v1/events": {
      "get": {
        "operationId": "stream_events_v1_events_get",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Server-sent events"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Role too low"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stream Events",
        "tags": [
          "events"
        ]
      }
    },
    "/v1/gateways": {
      "get": {
        "operationId": "list_gateways_v1_gateways_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gateways"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Role too low"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Gateways",
        "tags": [
          "gateways"
        ]
      }
    },
    "/v1/gateways/failover": {
      "post": {
        "operationId": "failover_v1_gateways_failover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailoverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failover"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Role too low"
          },
          "409": {
            "description": "No gateway is MASTER"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "description": "No agent client configured"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Failover",
        "tags": [
          "gateways"
        ]
      }
    },
    "/v1/ipam": {
      "get": {
        "operationId": "ipam_v1_ipam_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ipam"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Role too low"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ipam",
        "tags": [
          "pool"
        ]
      }
    },
...
    "/v1/pool": {
      "get": {
        "operationId": "pool_v1_pool_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pool"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Role too low"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pool",
        "tags": [
          "pool"
        ]
      }
    },
...
  }
}